pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kr-data-portal-client"
version = "0.1.8"
description = "High-quality Python client for KR Public Data Portal"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "aiolimiter>=1.1.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=3.6.0",
    "hatch>=1.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kr_data_portal"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "DTZ", "PT", "RET", "SIM", "PTH"]
ignore = ["N815", "N802", "N803", "E501", "PTH123"]

[tool.ruff.lint.isort]
known-first-party = ["kr_data_portal"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
